content-processor/requirements.txt
PyJWT==2.9.0
awslambdaric==2.2.1
bs4==0.0.2
email-validator==2.2.0
fastapi==0.114.1
fastapi-cli==0.0.5
httptools==0.6.1
langchain-anthropic==0.2.3
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-openai==0.2.4
mangum==0.19.0
nltk==3.9.1
pinecone==5.1.0
prisma==0.15.0
python-multipart==0.0.9
uvloop==0.20.0
voyageai==0.2.4
watchfiles==0.24.0
websockets==13.0.1
wheel==0.44.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.text==3.12.1
ordered-set==4.1.0
platformdirs==4.2.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
aiohttp==3.11.8
pytube==15.0.0
pytubefix==6.15.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
urllib3==2.2.3
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6
asyncpg==0.29.0
boto3==1.35.10
gitdb==4.0.11
GitPython==3.1.43
httpx==0.27.2
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pydantic==2.10.2
pydub==0.25.1
aiohttp==3.11.8
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
autocommand==2.2.2
awslambdaric==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
boto3==1.35.10
botocore==1.35.71
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.114.1
fastapi-cli==0.0.5
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==3.12.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-anthropic==0.2.3
langchain-community==0.3.0
langchain-core==0.3.21
langchain-groq==0.2.0
langchain-openai==0.2.4
langchain-text-splitters==0.3.2
langsmith==0.1.147
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.55.3
ordered-set==4.1.0
orjson==3.10.12
packaging==24.2
pillow==11.0.0
pinecone==5.1.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
prisma==0.15.0
propcache==0.2.0
pycryptodomex==3.21.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytubefix==6.15.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
s3transfer==0.10.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.38.6
StrEnum==0.4.15
tenacity==8.5.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typeguard==4.4.1
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.20.0
voyageai==0.2.4
watchfiles==0.24.0
websockets==13.0.1
yarl==1.18.0
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6
zipp==3.21.0
redis==5.2.0
deep-translator==1.11.4
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1